Beaer Token Class
Created a bearer Token class. Its main function is to return a URLRequest with a bearer token in its header.  And it return for either POST or GET requests

Added a 'RequestType' enum which will essentiall auto-validate the strings needed for 'httpMethod' of  a URL Request

Created a Protocol called 'Header' and extended it to provide abstraction, hiding the implementation needed for creating a URLRequest. (Bearer Token implements this protocol and only needs to use a '.request' property to recieve a fully initialised URL Request )